recordset: try to make r.Close kill r.Results chan
This attempts to fix a weakness in the current Recordset method set
(lack of mostly synchronous read) and a goroutine leak caused by
a forever-blocking send on the channel returned by Results(). The
blocking channel is only closed if and only if the entire record set is
read to completion despite it being close-able.

Currently, the results channel appears to be held open on sends because
Close() hits the cancellation channel but the Results() channel ignores
it, so it attempts to send on a channel that may have stopped being
received on and hangs.

This attempts to address that in two ways:

1. Add a synchronous Read method that does away with the channel,
   allowing consumers of the API to model concurrent access more
   without using the two to-be-unexported Records and Errors channels
   and without assuming that the one and only concurrent reader
   provided by the client library is adequate (since the current
   behavior of leaking goroutines after closing is bad).  This could
   obviously be renamed, but it seemed most appropriate to read
   a record from a record set.

2. Require all sends and receives from Records and Errors inside of the
   Results() reader goroutine to also receive on the cancelled channel,
   allowing the goroutine to bail out early if need be. Unlike Read(),
   this will not return an error (ErrRecordsetClosed) if the recordset
   is closed.

Close now returns an error if and only if it has been called more than
once. This is still safe to call more than once, but it provides an
indication that the record set received a double-close, which is useful
to know when debugging.

Previously, there were flush loops in the goroutine for handling what
I assume is an odd attempt to clean out a channel that will be GC'd
(along with its contents) after being closed in the Close() method.
This does not account for any other goroutines that may be holding the
waitgroup open far longer than needed. These flush loops have been to
not waste time on memory that will be scrapped.