don't share the onComplete channel
With multiple readers only 1 reader would have gotten the error.
Make the channel 1 long so the goroutine can always write to it and will
always finish.
Remove a useless 'select'.
No explicit `<- nil` needed, the channel will be closed on return.
Don't keep the channel internally, might be a race.