[CLIENT-2274] Use constant sized connection buffers
The client would use a single buffer on the connection and would grow it
per demand in case it needed a bigger buffer, but would not shrink it.
This helped with avoiding using buffer pools and the associated
synchrinization, but resulted in hugew memory use in case there were a
few large records in the results, even if they were infrequent.
This changeset does two things:
	1. Will use a memory pool for large records only. Large records
	   are defined as records bigger than aerospike.DefaultBufferSize.
	   This is a tiered pool with different buffer sizes. The pool
	   uses sync.Pool under the cover, releasing unused buffers back to the
	   runtime.

	2. By using bigger aerospike.DefaultBufferSize values, the user can
	   immitate the old behavior, so no memory pool is used.

This change should result in much lower memory use by the client.