Convert command.getNamespaces to return a Iter.Seq2 instead of a map
The map will forever require heap allocations. The Iterator on the other hand can be optimized away to NOT require allocations. At the moment Go's compiler is not that sophisticated and still allocates memory for the iterator, but I expect that to be resolved in the future. Having said that, Iterator still performs significantly better and allocates less memory even today.